move the state machine tests to their own module
